# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/subscription:
        post:
            tags:
                - Subscription
            operationId: Subscription_CreateSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.CreateSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.CreateSubscriptionReply'
    /v1/subscription/{uid}:
        get:
            tags:
                - Subscription
            operationId: Subscription_GetSubscription
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.SubscriptionItem'
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscription
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionReply'
        delete:
            tags:
                - Subscription
            operationId: Subscription_DeleteSubscription
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.DeleteSubscriptionReply'
    /v1/subscription/{uid}/alert-levels:
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscriptionAlertLevels
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionAlertLevelsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionAlertLevelsReply'
    /v1/subscription/{uid}/data-sources:
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscriptionDataSources
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionDataSourcesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionDataSourcesReply'
    /v1/subscription/{uid}/notify-types:
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscriptionNotifyTypes
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionNotifyTypesRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionNotifyTypesReply'
    /v1/subscription/{uid}/status:
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscriptionStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionStatusReply'
    /v1/subscription/{uid}/strategy-groups:
        put:
            tags:
                - Subscription
            operationId: Subscription_UpdateSubscriptionStrategyGroups
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionStrategyGroupsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.UpdateSubscriptionStrategyGroupsReply'
    /v1/subscriptions:
        get:
            tags:
                - Subscription
            operationId: Subscription_ListSubscription
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: namespaceUid
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.ListSubscriptionReply'
    /v1/subscriptions/select:
        get:
            tags:
                - Subscription
            operationId: Subscription_SelectSubscription
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: namespaceUid
                  in: query
                  schema:
                    type: string
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: nextUid
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sovereign.api.v1.SelectSubscriptionReply'
components:
    schemas:
...
        sovereign.api.v1.CreateSubscriptionReply:
            type: object
            properties: {}
        sovereign.api.v1.CreateSubscriptionRequest:
            type: object
            properties:
                userId:
                    type: string
                namespaceUid:
                    type: string
                type:
                    type: string
                name:
                    type: string
                description:
                    type: string
...
        sovereign.api.v1.DeleteSubscriptionReply:
            type: object
            properties: {}
...
        sovereign.api.v1.ListSubscriptionReply:
            type: object
            properties:
                subscriptions:
                    type: array
                    items:
                        $ref: '#/components/schemas/sovereign.api.v1.SubscriptionItem'
                total:
                    type: string
...
        sovereign.api.v1.SelectItem:
            type: object
            properties:
                value:
                    type: string
                label:
                    type: string
                disabled:
                    type: boolean
                tooltip:
                    type: string
            description: SelectItem 通用选择项消息
...
        sovereign.api.v1.SelectSubscriptionReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/sovereign.api.v1.SelectItem'
                total:
                    type: string
                nextUid:
                    type: string
                hasMore:
                    type: boolean
...
        sovereign.api.v1.SubscriptionItem:
            type: object
            properties:
                uid:
                    type: string
                userId:
                    type: string
                namespaceUid:
                    type: string
                type:
                    type: string
                name:
                    type: string
                description:
                    type: string
                status:
                    type: integer
                    format: enum
                strategyGroupUids:
                    type: array
                    items:
                        type: string
                datasourceUids:
                    type: array
                    items:
                        type: string
                alertLevels:
                    type: array
                    items:
                        type: integer
                        format: int32
                notifyTypes:
                    type: array
                    items:
                        type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
...
        sovereign.api.v1.UpdateSubscriptionAlertLevelsReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionAlertLevelsRequest:
            type: object
            properties:
                uid:
                    type: string
                alertLevels:
                    type: array
                    items:
                        type: integer
                        format: int32
        sovereign.api.v1.UpdateSubscriptionDataSourcesReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionDataSourcesRequest:
            type: object
            properties:
                uid:
                    type: string
                datasourceUids:
                    type: array
                    items:
                        type: string
        sovereign.api.v1.UpdateSubscriptionNotifyTypesReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionNotifyTypesRequest:
            type: object
            properties:
                uid:
                    type: string
                notifyTypes:
                    type: array
                    items:
                        type: string
        sovereign.api.v1.UpdateSubscriptionReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionRequest:
            type: object
            properties:
                uid:
                    type: string
                name:
                    type: string
                description:
                    type: string
        sovereign.api.v1.UpdateSubscriptionStatusReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
        sovereign.api.v1.UpdateSubscriptionStrategyGroupsReply:
            type: object
            properties: {}
        sovereign.api.v1.UpdateSubscriptionStrategyGroupsRequest:
            type: object
            properties:
                uid:
                    type: string
                strategyGroupUids:
                    type: array
                    items:
                        type: string
...